Convert description size variables to u8

+++ voting/src/error.rs
@@ -13,10 +13,10 @@ pub enum ContractError {
     ExcessiveWithdraw { max_amount: u128 },
 
     #[error("description too short (minimum description length {min_desc_length:?})")]
-    DescriptionTooShort { min_desc_length: usize },
+    DescriptionTooShort { min_desc_length: u8 },
 
     #[error("description too long (maximum description length {max_desc_length:?})")]
-    DescriptionTooLong { max_desc_length: usize },
+    DescriptionTooLong { max_desc_length: u8 },
 
     #[error("no stake")]
     PollNoStake {},
